Refactor template handler (#1117)
Template handlers use common code to handle rendering of jinja
and python to cloudformation.  This is a refactor to move that common
code into a helper moddule so that each handler would not need to
impolement it in their own custom handler code.
