+++ tests/test_template_handlers/test_file.py
@@ -1,9 +1,8 @@
 import os
-
 import pytest
 import yaml
 
-from sceptre.template_handlers.file import File
+import sceptre.template_handlers.helper as helper
 
 
 @pytest.mark.parametrize("filename,sceptre_user_data,expected", [
@@ -56,7 +55,7 @@ def test_render_jinja_template(filename, sceptre_user_data, expected):
         os.getcwd(),
         "tests/fixtures/templates"
     )
-    result = File._render_jinja_template(
+    result = helper.render_jinja_template(
         template_dir=jinja_template_dir,
         filename=filename,
         jinja_vars={"sceptre_user_data": sceptre_user_data}
